removing circular dependencies
