+++ packages/space-opera/src/components/animation_controls/animation_controls.ts
@@ -22,7 +22,7 @@ import '@polymer/paper-item';
 
 import {customElement, html, internalProperty, query} from 'lit-element';
 
-import {State} from '../../space_opera_base.js';
+import {State} from '../../types.js';
 import {dispatchAnimationName, dispatchAutoplayEnabled} from '../config/reducer';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 import {CheckboxElement} from '../shared/checkbox/checkbox.js';
