+++ packages/space-opera/src/components/camera_settings/components/radius_limits.ts
@@ -18,7 +18,7 @@
 
 import {customElement, internalProperty} from 'lit-element';
 
-import {State} from '../../../space_opera_base.js';
+import {State} from '../../../types.js';
 import {Camera} from '../camera_state.js';
 import {dispatchRadiusLimits} from '../reducer.js';
 import {Limits} from '../types.js';
