+++ packages/space-opera/src/components/ibl_selector/reducer.ts
@@ -17,8 +17,8 @@
 
 import {createSafeObjectUrlFromArrayBuffer} from '@google/model-viewer-editing-adapter/lib/util/create_object_url.js'
 
-import {Action, reduxStore} from '../../space_opera_base.js';
-
+import {reduxStore} from '../../space_opera_base.js';
+import {Action} from '../../types.js';
 import {INITIAL_ENVIRONMENT_IMAGES} from './initial_environment_images.js';
 import {EnvironmentImage} from './lighting_state.js';
