+++ packages/space-opera/src/components/inspector/inspector.ts
@@ -22,7 +22,7 @@
 import {createSafeObjectUrlFromUnsafe, SafeObjectUrl} from '@google/model-viewer-editing-adapter/lib/util/create_object_url.js'
 import {customElement, html, internalProperty, PropertyValues} from 'lit-element';
 
-import {State} from '../../space_opera_base.js';
+import {State} from '../../types.js';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 import {TexturesById} from '../materials_panel/material_state.js';
