+++ packages/space-opera/src/components/model_viewer_preview/reducer.ts
@@ -18,7 +18,8 @@
 import {GltfModel} from '@google/model-viewer-editing-adapter/lib/main.js'
 import {ModelViewerElement} from '@google/model-viewer/lib/model-viewer';
 
-import {Action, reduxStore} from '../../space_opera_base.js';
+import {reduxStore} from '../../space_opera_base.js';
+import {Action} from '../../types.js';
 import {dispatchSetEdits} from '../materials_panel/reducer.js';
 import {getGltfEdits, GltfEdits, INITIAL_GLTF_EDITS} from '../model_viewer_preview/gltf_edits.js';
