+++ packages/space-opera/src/components/model_viewer_snippet/components/download_button.ts
@@ -23,7 +23,7 @@ import {safeDownloadCallback} from '@google/model-viewer-editing-adapter/lib/uti
 import JSZip from 'jszip';
 import {css, customElement, html, internalProperty} from 'lit-element';
 
-import {State} from '../../../space_opera_base.js';
+import {State} from '../../../types.js';
 import {ConnectedLitElement} from '../../connected_lit_element/connected_lit_element.js';
 
 interface Payload {
